Renamed library
