+++ pyiqvia/errors.py
